cmd+loopd: move loopd to new folder loopd
